pyproject.toml
[project]
name = "keep-vibing"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pywinpty>=2.0.14",
    "pyjwt[crypto]>=2.11.0",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
